# packages/api-client/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ConsentRequestDetail": {
        "properties": {
          "appointment_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appointment Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ConsentRequestStatus"
          },
          "submission": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConsentSubmissionRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "template_version_id": {
            "format": "uuid",
            "title": "Template Version Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "patient_id",
          "appointment_id",
          "template_version_id",
          "status",
          "expires_at",
          "created_at"
        ],
        "title": "ConsentRequestDetail",
        "type": "object"
      },
      "ConsentRequestInvalidate": {
        "properties": {
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ConsentRequestInvalidate",
        "type": "object"
      },
      "ConsentRequestRead": {
        "properties": {
          "appointment_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appointment Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ConsentRequestStatus"
          },
          "template_version_id": {
            "format": "uuid",
            "title": "Template Version Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "patient_id",
          "appointment_id",
          "template_version_id",
          "status",
          "expires_at",
          "created_at"
        ],
        "title": "ConsentRequestRead",
        "type": "object"
      },
      "ConsentRequestStatus": {
        "enum": [
          "pending",
          "completed",
          "expired",
          "invalidated"
        ],
        "title": "ConsentRequestStatus",
        "type": "string"
      },
...
      "ConsentSubmissionResult": {
        "properties": {
          "eligibility_result": {
            "$ref": "#/components/schemas/EligibilityResult"
          },
          "submission_id": {
            "format": "uuid",
            "title": "Submission Id",
            "type": "string"
          }
        },
        "required": [
          "submission_id",
          "eligibility_result"
        ],
        "title": "ConsentSubmissionResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Medical Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/consents/requests": {
      "get": {
        "operationId": "list_consent_requests_api_v1_consents_requests_get",
        "parameters": [
          {
            "in": "query",
            "name": "patient_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ConsentRequestStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "needs_review",
            "required": false,
            "schema": {
              "default": false,
              "title": "Needs Review",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsentRequestRead"
                  },
                  "title": "Response List Consent Requests Api V1 Consents Requests Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Consent Requests",
        "tags": [
          "consents"
        ]
      },
      "post": {
        "description": "Creates a single-use consent link. Only the token hash is persisted \u2014\nthe raw token returned here is what gets embedded in the WhatsApp link.",
        "operationId": "create_consent_request_api_v1_consents_requests_post",
        "parameters": [
          {
            "in": "query",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "template_version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Template Version Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "appointment_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Create Consent Request Api V1 Consents Requests Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Consent Request",
        "tags": [
          "consents"
        ]
      }
    },
    "/api/v1/consents/requests/{request_id}": {
      "get": {
        "operationId": "get_consent_request_api_v1_consents_requests__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentRequestDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Consent Request",
        "tags": [
          "consents"
        ]
      }
    },
    "/api/v1/consents/requests/{request_id}/invalidate": {
      "post": {
        "operationId": "invalidate_consent_request_api_v1_consents_requests__request_id__invalidate_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentRequestInvalidate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentRequestRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Invalidate Consent Request",
        "tags": [
          "consents"
        ]
      }
    },
    "/api/v1/consents/requests/{request_id}/resend": {
      "post": {
        "description": "Same response shape as create_consent_request: only the raw token is\never returned over the wire, never persisted (see ConsentRequest's\ndocstring) \u2014 the caller embeds it in a fresh WhatsApp link.",
        "operationId": "resend_consent_request_api_v1_consents_requests__request_id__resend_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Resend Consent Request Api V1 Consents Requests  Request Id  Resend Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Resend Consent Request",
        "tags": [
          "consents"
        ]
      }
    },
    "/api/v1/consents/templates": {
      "get": {
        "operationId": "list_consent_templates_api_v1_consents_templates_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsentTemplateRead"
                  },
                  "title": "Response List Consent Templates Api V1 Consents Templates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Consent Templates",
        "tags": [
          "consents"
        ]
      },
      "post": {
        "operationId": "create_consent_template_api_v1_consents_templates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentTemplateCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentTemplateRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Consent Template",
        "tags": [
          "consents"
        ]
      }
    },
    "/api/v1/consents/templates/versions/{version_id}": {
      "get": {
        "description": "Staff-side full content of a version, independent of consent-request\nstatus \u2014 used both to preview a draft before publishing and to review\nthe exact questions a patient answered on a completed submission.",
        "operationId": "get_consent_template_version_api_v1_consents_templates_versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentTemplateVersionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Consent Template Version",
        "tags": [
          "consents"
        ]
      }
    },
    "/api/v1/consents/templates/{template_id}/versions/{version_id}/publish": {
      "post": {
        "operationId": "publish_consent_template_version_api_v1_consents_templates__template_id__versions__version_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Template Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentTemplateVersionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Publish Consent Template Version",
        "tags": [
          "consents"
        ]
      }
    },
...
    "/api/v1/public/consents/{token}": {
      "get": {
        "operationId": "get_consent_form_api_v1_public_consents__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentFormRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Consent Form",
        "tags": [
          "public",
          "public-consents"
        ]
      }
    },
    "/api/v1/public/consents/{token}/submit": {
      "post": {
        "operationId": "submit_consent_form_api_v1_public_consents__token__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentSubmissionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentSubmissionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Consent Form",
        "tags": [
          "public",
          "public-consents"
        ]
      }
    },
...
  }
}